requirements.txt
anthropic>=0.40.0
supabase>=2.10.0
twilio>=9.0.0
python-dotenv>=1.0.0
requests>=2.32.0
